backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
pydantic-settings>=2.2.1
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# Pin bcrypt below 4.x: passlib 1.7.x has a known compatibility break with bcrypt>=4.0
bcrypt>=3.2.0,<4.0

# Rate limiting
slowapi>=0.1.9

# Testing (required for CI and any pytest run)
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
